[REFACTOR] #4 - 관심사 분리
- 관심사 분리까지라고 하기는 좀 그럴 수 있지만, 프래그먼트 코드에 없어도 될 것 같은 코드는 분리함.
