--- app/src/main/java/com/ivyclub/contact/util/RecyclerViewBindingExtensions.kt
@@ -1,14 +1,0 @@
-package com.ivyclub.contact.util
-
-import android.view.LayoutInflater
-import android.view.ViewGroup
-import androidx.annotation.LayoutRes
-import androidx.databinding.DataBindingUtil
-import androidx.databinding.ViewDataBinding
-
-fun <T : ViewDataBinding> ViewGroup.binding(
-    @LayoutRes layoutRes: Int,
-    attachToParent: Boolean = false
-): T {
-    return DataBindingUtil.inflate(LayoutInflater.from(context), layoutRes, this, attachToParent)
-}
