+++ app/src/main/res/layout/fragment_friend.xml
@@ -72,7 +72,7 @@
             android:layout_width="0dp"
             android:layout_height="wrap_content"
             android:layout_marginHorizontal="20dp"
-            android:layout_marginTop="4dp"
+            android:layout_marginTop="8dp"
             android:background="@drawable/shape_search_friend_list_edit_text"
             android:hint="@string/fragment_friend_find_friend"
             android:importantForAutofill="no"
